Refactor to move logic into a hook.
